requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
pillow==10.1.0
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.4
gunicorn==21.2.0